CBP CSH BNK Updates

+++ CBP/plugin/stock_transactions/fermat-cbp-plugin-stock-transactions-crypto-money-destock-bitdubai/src/main/java/com/bitdubai/fermat_cbp_plugin/layer/stock_transactions/crypto_money_destock/developer/bitdubai/version_1/utils/CryptoTransactionParametersWrapper.java
@@ -4,6 +4,7 @@ import com.bitdubai.fermat_api.layer.all_definition.enums.CryptoCurrency;
 import com.bitdubai.fermat_api.layer.all_definition.enums.FiatCurrency;
 import com.bitdubai.fermat_ccp_api.layer.crypto_transaction.hold.interfaces.CryptoHoldTransactionParameters;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -14,7 +15,7 @@ public class CryptoTransactionParametersWrapper implements CryptoHoldTransaction
     private  CryptoCurrency  cryptoCurrency;
     private  String          walletPublicKey;
     private  String          publicActorKey;
-    private  float           amount;
+    private  BigDecimal      amount;
     private  String          memo;
     private  String          publicKeyPlugin;
 
@@ -22,7 +23,7 @@ public class CryptoTransactionParametersWrapper implements CryptoHoldTransaction
                                               CryptoCurrency cryptoCurrency,
                                               String walletPublicKey,
                                               String publicActorKey,
-                                              float amount,
+                                              BigDecimal amount,
                                               String memo,
                                               String publicKeyPlugin
     ){
@@ -78,12 +79,12 @@ public class CryptoTransactionParametersWrapper implements CryptoHoldTransaction
     }
 
     @Override
-    public float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
     @Override
-    public void setAmount(float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
